services/auth-service/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
pyjwt==2.8.0
passlib[bcrypt]==1.7.5
